requirements.txt
fastapi
uvicorn[standard]
numpy
scipy
pyjwt
passlib[bcrypt]
python-multipart
sqlalchemy